created test file and logic.js file relates #78" " ""
