+++ src/handlers.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const fs = require('fs');
+const findMatches = require('./logic')
 
 const handleHome = (response) => {
   const filePath = path.join(__dirname, "..", "public", "index.html");
@@ -13,32 +14,6 @@ const handleHome = (response) => {
     }
   });
 }
-
-const findMatches = (str, callback) => {
-  const filePath = path.join(__dirname, "..", "nameOfStar.txt");
-  var result = [];
-  fs.readFile(filePath, 'utf8', (err, file) => {
-    if (err) {
-      console.log(err);
-    } 
-    else {
-        file.toLowerCase().split(",").forEach((star) => { //split at the comma and loop through the string
-           if (star.startsWith(str) && result.length <= 10) {  //if string matches the first few inputs
-             result.push(star); //push the star name into the auto result
-           }
-         });
-         if (result.length <= 10) {
-           file.toLowerCase().split(",").forEach((star) => {
-             if ((star.indexOf(str) !== -1) && (result.length <= 10) && (result.indexOf(star) === -1)) {
-               result.push(star);
-             }
-           })
-          
-};
-       callback(result);
-     }
-   });
-  }
 
 const handleAuto = (request, response) => {
   const str = decodeURI(request.url.split('?')[1]);
@@ -77,7 +52,6 @@ const handleCSS = (request, response) => {
     }
   });
 }
-
 
 const handleImages = (response, url) => {
   const extension = url.split('.')[1];
